requirements.txt
# Core Django
Django>=4.2
gunicorn

# Database
psycopg2-binary
dj-database-url

# Configuration
python-decouple

# Static Files
whitenoise

# CORS (for API access if needed)
django-cors-headers

# HTTP Requests (if needed by games)
requests

# Caching
redis>=4.5.0
django-redis>=5.2.0

# Performance monitoring (optional)
# django-debug-toolbar  # For development only